moved routes and cleaned up
